Move query_reformulator to services directory
